api/requirements.txt
django
djangorestframework
pandas
django-cors-headers
djangorestframework-csv